# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DUNS Engine API for automated D-U-N-S acquisition and post-incorporation identity orchestration
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CaseStatus = typeof CaseStatus[keyof typeof CaseStatus];


export const CaseStatus = {
  pending_match: 'pending_match',
  match_found: 'match_found',
  queued: 'queued',
  submitted: 'submitted',
  received: 'received',
  in_progress: 'in_progress',
  closed_created: 'closed_created',
  closed_exists: 'closed_exists',
  closed_failed: 'closed_failed',
  challenged: 'challenged',
  error: 'error',
} as const;

export interface Case {
  id: number;
  /** Unique case identifier (UUID) */
  case_id: string;
  status: CaseStatus;
  /** 9-digit D-U-N-S number, populated when resolved */
  duns_number?: string | null;
  business_name: string;
  street_address: string;
  city: string;
  state?: string | null;
  postal_code?: string | null;
  /** ISO 3166-1 alpha-2 country code */
  country_code: string;
  phone?: string | null;
  requestor_email?: string | null;
  /** D&B IDR confidence score (0-10). Score >= 8 means entity already has a DUNS. */
  match_confidence?: number | null;
  /** Resolution detail for closed_failed cases */
  resolution_code?: string | null;
  /** D&B internal case ID for submitted cases */
  db_case_id?: string | null;
  created_at: string;
  updated_at: string;
  closed_at?: string | null;
  /** Number of processing attempts made by the durable worker */
  attempts?: number;
  /** Most recent error recorded by the worker (set for retried/failed cases) */
  last_error?: string | null;
  /** When the durable worker will next act on this case */
  next_action_at?: string | null;
}

export interface CaseListResponse {
  cases: Case[];
  total: number;
  limit: number;
  offset: number;
}

export interface CreateCaseRequest {
  /** Legal entity name */
  business_name: string;
  street_address: string;
  city: string;
  state?: string | null;
  postal_code?: string | null;
  /** ISO 3166-1 alpha-2 country code */
  country_code?: string;
  phone?: string | null;
  /** Email for resolution notifications */
  requestor_email?: string | null;
}

export interface Metrics {
  total_requests: number;
  pending: number;
  submitted: number;
  in_progress: number;
  /** Cases resolved with a new DUNS number */
  completed_created: number;
  /** Cases where DUNS already existed (match found) */
  completed_exists: number;
  failed: number;
  /** Average hours from submission to resolution */
  avg_resolution_hours?: number | null;
  /** Percentage of non-failed closed cases */
  success_rate: number;
}

export interface QuotaEntry {
  country_code: string;
  /** D&B research type (Mini, Standard, etc.) */
  research_type: string;
  date: string;
  used: number;
  limit: number;
  remaining: number;
}

export interface AtlasIncomingPayload {
  /** Stripe Atlas company ID */
  company_id: string;
  legal_name: string;
  street_address: string;
  city: string;
  state?: string | null;
  postal_code?: string | null;
  country_code?: string;
  founder_email?: string | null;
}

export interface AtlasAcknowledgement {
  case_id: string;
  status: CaseStatus;
  message: string;
}

export interface RunDemoRequest {
  /**
     * A 9-digit DUNS number to deliver at the end of the live run.
     * @pattern ^[0-9]{9}$
     */
  duns_number: string;
  /** Optional business name to display on the case. Defaults to a demo placeholder. */
  business_name?: string;
}

export interface DemoResetResponse {
  ok: boolean;
  cases_seeded: number;
  quota_seeded: number;
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetCasesParams = {
status?: CaseStatus;
/**
 * @minimum 1
 * @maximum 200
 */
limit?: number;
/**
 * @minimum 0
 */
offset?: number;
};