# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DUNS Engine API for automated D-U-N-S acquisition and post-incorporation identity orchestration
 * OpenAPI spec version: 0.1.0
 */
import type { CaseStatus } from './caseStatus';

export interface Case {
  id: number;
  /** Unique case identifier (UUID) */
  case_id: string;
  status: CaseStatus;
  /** 9-digit D-U-N-S number, populated when resolved */
  duns_number?: string | null;
  business_name: string;
  street_address: string;
  city: string;
  state?: string | null;
  postal_code?: string | null;
  /** ISO 3166-1 alpha-2 country code */
  country_code: string;
  phone?: string | null;
  requestor_email?: string | null;
  /** D&B IDR confidence score (0-10). Score >= 8 means entity already has a DUNS. */
  match_confidence?: number | null;
  /** Resolution detail for closed_failed cases */
  resolution_code?: string | null;
  /** D&B internal case ID for submitted cases */
  db_case_id?: string | null;
  created_at: Date;
  updated_at: Date;
  closed_at?: Date | null;
  /** Number of processing attempts made by the durable worker */
  attempts?: number;
  /** Most recent error recorded by the worker (set for retried/failed cases) */
  last_error?: string | null;
  /** When the durable worker will next act on this case */
  next_action_at?: Date | null;
}