# lib/api-zod/src/generated/types/caseListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DUNS Engine API for automated D-U-N-S acquisition and post-incorporation identity orchestration
 * OpenAPI spec version: 0.1.0
 */
import type { Case } from './case';

export interface CaseListResponse {
  cases: Case[];
  total: number;
  limit: number;
  offset: number;
}