# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DUNS Engine API for automated D-U-N-S acquisition and post-incorporation identity orchestration
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}